% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{graph for soil science lab}
\usage{
boxplot(data, x, y, xlab, ylab, fill, col)
}
\arguments{
\item{data}{use data}

\item{x}{x axis}

\item{y}{y axis}

\item{xlab}{x axis title name}

\item{ylab}{y axis title name}

\item{fill}{fill name You do not have to enter it if you do not need id}

\item{col}{color code}
}
\value{
graph
}
\description{
\code{boxplot} create boxplot
}
\examples{
# boxplot(data = soil data, x = Treatment, y = CUE)
}
