% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssl.NMDS.R
\name{f.envfit.analysis}
\alias{f.envfit.analysis}
\title{Envfit analysis}
\usage{
f.envfit.analysis(data, env)
}
\arguments{
\item{data}{Enter the "nmds.result" obtained from "f.NMDS.analysis}

\item{env}{Environment data. You read Environmental data sheet.}
}
\value{
env.vec
}
\description{
\code{f.envfit.analysis} Envfit analysis
}
\examples{
# environment <- read.csv(environment.csv, header = T, stringAsFactor = T)
#f.envfit.analysis(data = nmds.resuslt, env = environment)

}
