% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssl.NMDS.R
\name{f.NMDS.analysis}
\alias{f.NMDS.analysis}
\title{NMDS analysis}
\usage{
f.NMDS.analysis(data, loc)
}
\arguments{
\item{data}{use data}

\item{loc}{location data}
}
\value{
nmds.result

bata.dist
}
\description{
\code{f.NMDS.analysis} NMDS analysis
}
\examples{
# f.NMDS.analysis(data = input, loc = loc)
}
