% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssl.NMDS.R
\name{f.NMDS.graph}
\alias{f.NMDS.graph}
\title{NMDS graph}
\usage{
f.NMDS.graph(data = nmds.result, color, shape)
}
\arguments{
\item{data}{Enter the "data.score" obtained from "f.nmds.result"}

\item{color}{Enter the color name or hexadecimal color code. You can omit.}

\item{shape}{Enter the shape number. You can omit.}
}
\value{
NMDS.graph
}
\description{
\code{f.NMDS.graph} create NMDS plot
}
\examples{
# NMDS.graph(data = data.score)
# NMDS.graph(data = data.score, color = c("#0000ff", "#ffa500), shape = c(1, 2))
# NMDS.graph(data = data.score, env = env.vec)

}
