% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssl.NMDS.R
\name{f.NMDS.analysis}
\alias{f.NMDS.analysis}
\title{NMDS analysis}
\usage{
f.NMDS.analysis(data, dttype, subsamplesize, loc)
}
\arguments{
\item{data}{Amplicon or PLFA data}

\item{dttype}{You choice "A" or "P". If you analyse amplicon data, you choice "A". If you analyse PLFA data, you choice "P".}

\item{subsamplesize}{You enter sabsamplesize}

\item{loc}{location data. You read location data sheet.}
}
\value{
data.score

nmds.result
}
\description{
\code{f.NMDS.analysis} NMDS analysis
}
\examples{
# f.NMDS.analysis(data = input, dytype = "A", subsamplesize = 20000, loc = loc)
}
